backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pdfplumber==0.10.3
faiss-cpu==1.7.4
langchain==0.1.9
langchain-text-splitters==0.0.1
numpy==1.24.3
openai>=1.51.0
httpx>=0.27.2